Don't redocument on build
This somehow has worked:
http://teamcity.montagu.dide.ic.ac.uk:8111/viewLog.html?buildId=26494&buildTypeId=montagu_Orderly_BuildContainer&tab=buildLog&_focus=502#_state=502

but I don't actually see how. Because we run the container with
user permissions but copy the sources in with root, the document
step that pkgdown runs can't possibly succeed.

I am also opportunistically clearing out the docs directory before
build, which will mean that deleted files are not included