New approach to changelog that avoids json entirely
Now we store the changelog like everything else in the run
rds which means that they're easier to migrate etc.

The big downside of this is that we lose the "natural ordering"
of the serial key that was used, which preserved the ordering
of the original log.  Things are still orderable by the
report version though, so this affects only individual reports
with multiple changelog entries