Avoid including reference data in distribution
This moves the reference data out of the package by default, but
includes them in tests on travis.  Tests requiring reference data
(mostly migration tests) will be skipped over when not present.

On travis, the reference data will be available and it will be
an error if they are not present, rather than a skip.  On appveyor
the reference data is not available so we'll know that the skip
works.

Removing the data from the package should reduce the size by about
5MB which is large relative to most R package source sizes.